moved code related to running external CLI applications to its own python file.
